SCUBA2: add initial fakemap processing here (turned off by default)
I forgot that the fakemap recipe parameters should be added in
addition to any other reduce_scan parameters, so just add them to the
list of parameters to verify and then call _FAKEMAP_STEER_ to set
those parameters in the uhdr.

A new argument has been added to allow the user to select fakemap
parameter processing.