Regenerated sun230 and sun233 for hawaiki release
with manual removal of trailing space to allow merge from hawaiki
branch. This patch was not applied to the repository at the time
of the hawaiki release so the documentation is not up-to-date
for namaka.

Cherry-picked from 24bb31eb998d1ae3afd8ce9a2d1b869d7987e72f
on hawaiki