heterodyne: Divide the input files into chunks to prevent timeout in MAKECUBE.
Given a typical number of large time-series files, the MAKECUBE step
combining all of these requires vast resources to process and could
time out in ORAC-DR.  This divides the input files into chunks and
forms the tiled cubes for these chunks.  Files in each chunk are assigned
to the tiles that a full MAKECUBE would create.  The data are reassembled
averaging with CCDPACK:MAKEMOS, trimmed to the tile bounds.

There is a new argument called CHUNK to control the maximum number of
time-series files to process simultaneously in MAKECUBE, currently
defaulted to 5.

This does not correctly handle the SMURF extension NDFs.  It also does
not make any special provision for potential smoothing artefacts.