heterodyne: Correct file naming to mimic MAKECUBE behaviour.
I'd forgotten that the output file name changes if there is only
one output cube.

Note this still chunks on small datasets as it uses the number of
input files.  In Stare mode this is not suitable and a revised version
that uses the total file size is in development.