lib: Rewrite "-loop task"
The previously implemented "task" loop mode was very fragile since it
relied on "getting" parameters from the remote tasks via a poll whenever
the pipeline had finished processing the previous data. This was
complicated and prone to dropping frames unnecessarily and also had
problems when the frame counting in the remote tasks got unexpected
values.

This rewrite does things in a much more robust manner. There is now a
SCUBA-2 specific DRAMA task called "qlgather" that sets up proper DRAMA
monitors on the remote tasks. These monitors then update the values in
the tasks whenever the remote updates and when all parameters have been
received from all the active tasks the data are written to disk. The
next time data turn up the previous flag file is moved to a safe place
and the old files deleted. The new flag file is then written for the
pipeline.

The pipeline looks for a single flag file. When it finds it the flag
file is moved into a safe place and the files read by the pipeline. When
the loop is entered again the files from the previous run are deleted.

The new scheme is very responsive to changes and does not get confused
when remote tasks die and are restarted.