SCUBA2: the input map may be a Gaussian of given FWHM at map centre
The user specifies the FWHM in arcsec as FAKEMAP_FWHM in the recipe
parameter file, instead of FAKEMAP_MAP.

The amplitude of the Gaussian is scaled by the FAKESCALE parameter
used by makemap. The default amplitudes are equivalent to 1/4 Jy at
850/450 um respectively assuming standard FCFs.

It's probably more efficient to create one manually and pass it in for
large data sets, but this allows the parameter space to be explored
more easily.