lib: Rewrite to work for generic dimensionality
The previous implementation only worked for 2 or 3-D WCS and
then had lots of special-cased code to handle it. This version
will locate the SpecFrame and SkyFrame from an N-D WCS.

Additionally the bounds are now calculated for the pixel
centre and not the pixel corner. This matches the definition
used by SMURF MAKECUBE and MAKEMAP.