SCUBA2: add basic support for CUPID FindClumps
At the moment only Clumpfind and Fellwalker are supported by default,
unless the user provides their own config file.

Testing was done with SASSy data, and in order to match what's in the
paper, there are a couple of fudge factors in place to produce
catalogues with (more or less) the same s/n threshold as given by the
DETECT_SIGMA argument. Not very satisfactory but it will do to begin
with.