Add error handler for Starlink::HDSPACK
When we got errors from Starlink::HDSPACK routines we only got
an error status and none of the messages so the pipeline would
abort and no-one would understand why. Now we register a callback
with Starlink::HDSPACK so that orac_err can be called
appropriately.

This commit requires an update to Starlink::HDSPACK version 2.01