lib: Read the FITS header if files() is called for the first time
Some code in the pipeline did:

  $tmpFrm = $Frm->new;
  $tmpFrm->files( @files );

and expected hdr() to be populated. This was broken when I fixed
header syncing so that it would not immediately write back to raw
data. The headers were being read as a side effect of calling
sync_headers(). Now the files() method will automatically read
the header if this is the first time files have been stored.