SCUBA2: meta-primitive to calculate FCFs in the same way as PICARD recipe
Experience with S2SRO data and the PICARD recipe SCUBA2_FCFNEFD has
shown that the most stable FCF estimates are made with trimmed and
background-subtracted images. Therefore this primitive should be
included in recipes where FCF calculation is likely.

Currently the default settings match those of the PICARD recipe but
will undoubtedly change with further experience.