SCUBA2: write units of peak power measurement, optional log file with fit parameters
A new log file can be written, called log.fit_fcf which contains the
parameters of the fits carried out to estimate the BEAM and BEAMMATCH
FCFs. At this time, the major and minor axes are written along with
the peak amplitude (and its uncertainty), the fitted gamma (and
uncertainty) for the BEAM and BEAMMATCH fits respectively.