SCUBA2: use a smaller default fitting area for pointing observations
Pointing observations are usually cropped to an area as small as 60x60
arcsec**2, so the default fitting box size of 25 pixels is larger than
the map. In this case the primitive does not attempt to fit the source
as the logic assumes that an image with fewer good pixels than
contained in the fitting area is too poor to use.