heterodyne: Chunk assignment to tiles is much simpler.
A great simplification and efficiency gain is achieved, because we
have a MAKECUBE setup where the chunks' tile numbering and bounds
follow that of the full set of input data.  We use MAKECUBE TRIM=F
AUTOGRID=F and set new reference-pixel (REFPIX1, REFPIX2) parameters
once the output cube's geometry is determined.  Thus it is now merely
a question of collating by the index of the chunks' tiles (as was
expected in the original code).  This fixes a long-standing problem
causing a four- or five-pixel shift evident in comparisons with the
unchunked data, where the edges of tiles showed differences and strong
sources appeared with bas-relief shadows.

Also there's no need to mask borders as we set the tile border width
to zero.  MAKECUBE had a special case, annexing into adjacent tiles
if the edge tile had only data within the tile border region.  This
undocumented feature caused considerable bewilderment, and
misbehaviour in this primitive.