SCUBA2: store the map peak position for later retrieval
Having gone to the trouble of running the map through the matched
filter and fitting the source position, it might save some time to
store the map peak position in case it comes in handy at a later point
in the recipe. At the moment, just store an array reference with the
coordinates of the map peak.

Of course, there is plenty of scope for this to backfire as there is
no information on which map was used etc...