heterodyne: Remove intermediate chunk files.
Since the chunking was changed from file numbers to total size, the
chunking division can change between iterations.  Combine that with
using tile index to reassemble the chunks and using file existence to
see if there's a tile component in a given chunk, it is vital to
remove these files to avoid forming the mosaic using some data from an
earlier iteration.  The plan is to access the MAKECUBE OUTFILES output
parameter to obviate the need for the file-existence test.