heterodyne: Use ORAC::TempFile for OUTFILES.
In addition it uses File::Basename to exclude the long path of the
temporary file, as switching from a $$ temporary file with an
ORAC::TempFile increases the MAKECUBE command-line length beyond the
MSP limit of 512.  This is also applied to the text file supplying the
list of input NDFs for MAKECUBE.  These changes should eliminate the
risk of exceeding the maximum command-line length, and some earlier
truncations can be undone.