SCUBA2: use a temporary frame to define output file names
When using the global Frm to define the output file names, the raw
fastflat files end up on the intermediates array and so get deleted at
the end of the recipe. This is a Bad Thing(TM) since they are needed
again for flatfielding the next time round. Since the Frame is only
being used to define the output file names, using a temporary object
will do just fine.