SCUBA2: let user know if the source in the given image was extended
The warning message about not locating a point source is a little
misleading - the subtle distinction between locating a source and
locating a *point* source is easy to miss. The immediate
interpretation tends to be that there is *no* source at all in the
image. Therefore, when no *point* source could be found, check to see
if the source was extended and let the user know that.