SCUBA2: calculate an effective NEFD by factoring out the number of bolometers
Calculate an effective NEFD from NEFD / sqrt(Nbol). This can be
calculated for both cases (noise and map data), and is written to the
log file along with the number of bolometers used.

Note that the effective NEFD is not calculated for co-added (ie Group)
maps as the number of bolometers is not known.