SCUBA2: calculate NEFDs for Frame images in summit and ql pipelines
The performance hit is negligible compared with the variation in
recipe run time so the online pipelines might as well calculate the
NEFD and report it. Note that the calculation is only done on the
current Frame image, not the Group image. This can always be changed at
a later date if required.