SCUBA2: reset Grp uhdr flags if a new observation has started
Subsequent observations of the same source were not being processed
because the percentage-completion flag was still set from the previous
observation. Use a new Grp uhdr flag to store the observation number
for the current map and if it changes reset all the relevant Grp uhdr
flags so that counting begins again.