SCUBA2: for offline processing, proposed and current aren't really applicable
When running the REDUCE_FASTFLAT recipe, the concept of `proposed' and
`current' solutions doesn't make a lot of sense. Instead, the
responsivity mosaics are given the titles First and Last. However, the
titles for the histograms cannot be changed, so I've still included
the proposed/current wording in the title so that they can be matched
with the mosaics.