SCUBA2: always make a copy of the raw data before proceeding
Recent changes in the pipeline made to accommodate fastflat processing
broke the recipe for discrete flatfields. The new Grp method
sort_by_subarray was failing because it ultimately requires write
access to the data (via sync_headers). The data were being copied
anyway for each heater setting in the QL and summit pipelines prior to
the last, so it only adds one extra copy call.

This means that _CALCULATE_FLATFIELD_ no longer needs to copy the data
(which was never done for fastflats anyway as they are always
processed when they turn up). See the next commit or two...