SCUBA2: use results retrieved from the calibration system if they exist
Previously the fastflat comparison relied entirely on storing a hash
of the flatfield solutions indexed by subarray in the Grp uhdr. This
works fine comparing the results of the flatfields within a *single*
observation, but doesn't allow the changes to be tracked *across*
observations.

The previous fastflat solutions are retrieved, and though a hash
reference is still used to communicate the results to downstream
primitives, it's all done within a block where Grp is localized so
there is never any issue of relying on the top-level Grp uhdr.

Tested on and seems to work with SETUPs, regular ql/summit
observations, discrete flatfields and with the REDUCE_FASTFLAT
recipe. The only slightly weird behaviour happens when running
REDUCE_FASTFLAT on focus observations.