SCUBA2: derive the RMS uncertainty and pass to findback
With some images (usually at 450 um) the data vcalues end up very high
(1e31 or more), and findback gives an error when calculating the RMS
itself. To workaround this feature, derive the RMS and pass it to
findback if it could be calculated (and checking that the returned
value of sigma is +ve). Seems to work and allows findback to deal with
these odd images.