SCUBA2: store responsivity mosaic as main DATA component in output file
It's confusing to have a responsivity mosaic file that can't be
displayed simply by passing it to, say, Gaia, so store the mosaic as
the main data component in the output file as well as an NDF
component. It's a bit wasteful to have the same data repeated in a
file, but it's not a lot of data and it keeps everybody happy - right?