SCUBA2: proceed with creating a cube if the current SMU position is the last
Assuming that all focus observations step the SMU position
symmetrically about a nominal zero point of 0.0 mm, we can use the
number of postions and the step to predict what the last SMU position
should be. If the current data are for that last position then cube
creation can proceed. This will allow the pipeline to deal with missed
data from previous SMU positions, though of course it can't deal with
the last position itself being skipped.