SCUBA2: initialize all possible amplitudes to -1, make sure midpoint index is an integer
My previous method for initializing the amps hash to -1 for skipped
files didn't actually work, so do that first with a map call.

If 1 position was skipped $midpt ends up half of an odd number which
doesn't translate to an array index very well... so just make it an int.