Calib: add a lookup table for peak flux densities for SCUBA-2
Up to now it's always been assumed that secondary calibrators are
point sources. This is not the case for CRL2688 so it's necessary to
allow the correct flux be chosen - the simplest way to do this is to
have a separate hash with the peak fluxes, and use the $ismap variable
to determine which hash to return to the caller.