SCUBA2: add capability to choose between peak/total fluxes for user fluxes
Use the capaibility of the recipe parameter parser to return an array
reference for multiple values to use a little code ref to return the
correct flux for the current value of $ismap.

Note: it is assumed that the first element of the array ref is the
peak flux, and the second element is the total.