Inst: add -grptrans for the JCMT QL pipelines, -nobatch for QL/SUMMIT
The -batch override causes problems with offline testing of the QL and
SUMMIT pipelines, so force -nobatch for all drip-fed JCMT pipelines.

The QL pipelines create a new Group for every observation, so there is
no need to keep old Groups and they should therefore be regarded as
transient (-grptrans).