SCUBA2: treat SUMMIT pipeline the same as the others
The original plan was to have the summit pipeline use all of the data
to derive the noise properties, but it seems that the method I
implemented doesn't work. As a quick fix (and so that the summit
pipeline doesn't fall behind too much) just process files in the
current Frame only, as is the case for the QL and offline pipelines.