SCUBA2: add options which will speed up calculation of average power spectra
Two modifications to this primitive bring the processing time for 30s
of data from 4 subarrays from 35 s to 15 s. This is achieved by:

1) Letting sc2clean handle the concatenation;
2) Allowing only a fraction of the data to be used in the calculation.

The new parameter NSAMP is the number of samples to use in each
file. The default is to use all the data.

However, note that NSAMP should ONLY be used for the drip-fed
pipelines - using it in offline mode where there are multiple files to
concatenate will probably break things horribly.