SCUBA2: display noise images as a focal-plane mosaic
Major changes to this primitive to create a focal-plane mosaic and
display that in a single window, rather than one window per
subarray. This has two advantages:
  1) it reduced the number of Kapview display windows
  2) it allows the relative noise of each suabrray to be seen

One further improvement is that the thresholding now takes place only
once on the mosaic, rather than on each subarray. Statistics are still
reported for each individual subarray.