SCUBA2: tell user how files are going into the average, use different suffix
The user is now informed of how many files are going into each
average, though it is assumed that the number of files is the same in
each Frame object.

Use a new suffix for the average noise files since they're not really
needed - this allows them to be deleted at the end of processing.

Don't strip the subscan from the name of the average noise files as it
causes problems with downstream naming of the average noise mosaic.