SCUBA2: add support for defining an annulus for background estimation
An annulus may now be used to estimate the background in an image. The
primitive takes a new argument - ANNULUS - which is a flag. The
default is to use the existing method (ie use the statistics of the
entire image outside of the signal aperture).

Additional arguments are INNER and OUTER which are the inner and outer
radii of he annulus, specified as multiples of the aperture
radius. Defaults are 1.25 and 2.0 respectively.