PICARD: new recipe and primitive to do basic aperture photometry on SCUBA-2 images
Perform aperture photometry on SCUBA-2 images using the chosen
method. By default (i.e. with no recipe parameters), this recipe will
calculate the flux within a 60-arcsec diameter aperture, corrected for
any DC offsets by using the rest of the image to estimate the
background. Alternatively the background may be estimated from an
annulus centred on the source. Finally, the user may request that the
autophotom packaged be used. Note, however, in that case that no
uncertainties are returned.

The results are written to a log file called log.flux.

Recipe parameters:

 * ANNULUS - flag to denote whether to use an annulus for background estimate
 * ANNULUS_INNER - inner radius of annulus
 * ANNULUS_OUTER - outer radius of annulus
 * APERTURE_RADIUS - radius of aperture
 * AUTOPHOTOM - flag to denote whether autophotom should be used
 * LOGFILE - flag to denote whether to write out a log file