SCUBA2: new recipe to derive the photon noise contribution
The dark noise and NEPs are calculated using the initial dark datafiles.

The sky noise and NEPs are calculated using the first non-ramp
open-shutter data. The same number of samples is used as for the dark
data, though this may be overridden with a recipe parameter.

A photon NEP file with suffix _phnep is written for each subarray.

Results are written to a log file, log.photnoise.