SCUBA2: make sure the current Frm is stored for retrieving the raw fastflats
Commit 46b62115b0 introduced a bug whereby the QL/online pipeline no
longer stored the fastflats in the calibration system. The raw
fastflat files are retrieved with a "map" call iterating over the
array of Frame objects. Unfortunately, the online modes this array was
empty whereas it should at least contain the current Frame
object. This is now the case so @Frms always has at least 1 Frame
object.