heterodyne: Do not open temporary files when each merely generates an NDF name.
This should fix the occasional issue where the ACSIS pipeline fails
because there are too many open files.  It occurred when the spectra
were divided into >~40 continuum sections because of the presence of a
similar number of emission lines.

Also removed a temporary file that was not being used that exacerbated
the problem.