SCUBA2: new primitive to determine if bolometers go out of range
The fastramp defines the range of values over which the flatfield
solution is valid. If any of the data files have bolometers which
exceed that range then they cannot be trusted in the map-maker. This
primitive determines the number of bolometers which go out of range
and writes entries to a log file called log.bolorange.