heterodyne: Handle case where all the observations are bad.
Instead of exiting when the recipe discovers all the observations
are bad, continue to process to create the OBS products.  This is
achieved by making a temporary group where the bad Frames are flagged
as good, proceeding as normal with that temporary group, but excluding
the final group processing instead removing group products.

This addresses the "log file ends with error status = -7" fault.