SCUBA2: skip analysis and display of data with no good bolometers.
The Frame uhdr should contain the hash reference NEPGOODBOL with
values returned from calcnoise indicating how many good bolometers
contributed to the NEP image. This primitive can retrieve that hash
reference and use it to determine whether or not it is worth
proceeding with analysis. We have to make that check here because we
want the modal noise value, which means we have to use histat, which
exits with an error if there are no good values in the array. (Stats,
on the other hand, is fine with a blank image.)