SCUBA2: deal with the cases where we don't have all four arrays
The attempt to speed up making pointing and focus maps by using 2 or
fewer arrays only worked for the cases when the INSTAP was defined or
all four arrays were present. Some observations have only 3 arrays, so
use the second and last subarray in the list. If fewer than 3 arrays
are present then just all of them.