SCUBA2: make default prefix for flag files the one that is used most often
I had originally set the default prefix to be the current subarray,
but as far as I can tell the SCUBA-2 pipeline never writes a flag file
that begins with the subarray, and every call to _WRITE_FLAG_FILE_ was
specifying PREFIX=s. So, I figured I might as well make that the
default instead.