heterodyne: Preserve reference pixel for the Frame.
After masking spectra the reference pixel from makecube autogrid can
change such that the bounds are shifted leading to a recipe crash for
single-spectrum cubes.  This fix stores the reference pixel for each
Frame in its internal header.  The ugly complication is the first time
the reference pixel is found it is for each subsystem through a
temporary Frame object, thus must be exported to the permanent Frame
object in _CREATE_CUBE_PER_SUBSYSTEM_.