JCMT: some SCUBA-2 Frames will have subheaders for the base position
In a small number of cases involving SCUBA-2 data, checking the hdr
for the base position will fail because of the presence of subheaders
- the Frame object can have data where the base position is not
defined for all files. Since a base position might exist in the
subheaders, check the first one (in line with choosing the current
file in the call to find_base_position) and use that if it is not
blank.