SCUBA2: new primitive to calculate the difference between the actual and demand azimuth values
This primitive reads the JCMTSTATE and calculates the RMS difference
between the actual and demand azimuth and elevation. The results are
reported to the screen and written to a log file, log.posdiff, for
monitoring by the StripChart.

The log file also contains the current az, el and min/max values of
the az and el differences, preserving the sign of the offset (so
minimum values are negative) to indicate whether the telescope is
ahead or behind.