jcmt: Scan all subheaders looking for valid base position
If TRACKSYS is present we don't actually know if BASEC1 will be
defined in the first subheader for SCUBA-2. Darks and fastflats
do not involve the TCS so BASECx is not defined. To overcome this
problem we scan through all the subheaders until we find defined
coordinates.

Without this we end up with an incorrect ASN_ID.