SCUBA2: compute noise properties rather than make maps
The QL pipeline is now primarily a quality assurance (QA) pipeline,
and the emphasis has shifted from producing images every 30 seconds to
tracking the noise performance of the arrays. This recipe has been
modified to process the science data as though they were noise
observations, and the results are written to log and index files for
monitoring by the stripchart.