Add -groupid for overriding the group ID string
This can be used to force a particular string to be used for
grouping all the files. This string is also used for the
ASN_ID if the method is supported by the frame class so that
externally supplied ASN_ID values can be written to JSA
output headers.

-onegroup still forces the files to be part of the same group
but does not change the group ID itself (or by extension the
ASN_ID).