scuba2: Use the model to find the peak
The image will have noise in it so we sometimes end up plotting
a location of a spike. Now we locate the peak location using the
model and only use the image to determine the min value for plotting.
(cherry picked from commit 2eb1c71237a7da622184d86f2d48945543c4114a)