scuba2: Warning about fitted peak should take background into account
BEAMFIT returns a source amplitude and a value for the background.
When comparing peak of source with amplitude the background must
be added back.

This was obvious for non-calibrator observations with the matched
filter that drive the background negative (the matched filter
fits should always force the bacground to zero but that is a
different problem).