scuba2: Add new secondary calibrators
The source fluxes are fairly uncertain but it's better to have them
in the code now.

Note that we also add "HL Tau b" for historical reasons (it's
actually a few arcsec to the left of HL Tau but for the purposes
of FCF calculation this is really a HL Tau observation).

The real fix is to define STC regions for each calibrator and
compare them with the observed map area.