PICARD: use autophotom if that recipe parameter is given
The recipe parameter AUTOPHOTOM is used to tell PICARD to calculate
FCFs using the autophotom routine. However, the flux reported later in
the SCUBA2_FCFNEFD recipe is not derived using autophotom, which users
might think strange if they've specified that recipe parameter. Now
the AUTOPHOTOM parameter can be used to force PICARD to report the
flux and its error using autophotom.