scuba2: remove 50-bolometer limit for matching flat
Processed flats are rarely if ever used for anything other than
assessing the flatfield solution itself, so the condition that a
suitable flatfield must have more than 50 bolometers is no longer
necessary. Furthermore it actually gets in the way of writing out all
the information necessary in a flag file during a SETUP observation.