SCUBA2: allow matching entries for all subarrays to be returned
Sometimes (e.g. SETUP observations) it is desireable to return all
possible flatfield solutions, regardless of whether or not they
succeeded. The ALLSUB parameter, if true, will loop over all possible
subarrays for the current wavelength and retrieve the relevant
flatfields if they exist. If false, only the subarrays for which data
exist in the current Frame will be matched.