SCUBA2: trap status from call to calcnoise
Check the return status from calcnoise and only store results if it
ran with no errors. The most common bad return status will be
SMF__BADFLAT for a failed flatfield, but for now it's not clear
whether that error condition should be handled in any other way than
for any other error, so proceed with using the good results if
available.