SCUBA2: store FCF uncertainties in primitive hash
The FCFs are stored in the primitive hash with the key FCF, but until
now the uncertainties in those FCFs were not.

Also make the use of AUTOPHOTOM controllable with a primitive argument
and not just a recipe parameter.
(cherry picked from commit ac25107d2acfceab989678005019c00746bf5c6d)