scuba2: add a QL pointing config file
This is a modified version of the standard SMURF pointing config file
but with fewer iterations (5 instead of 40). While the map may not
have formally converged within that number, it is good enough for
determining the peak position.
(cherry picked from commit e69608618405467f2f9fc4a7327f4b36615078ea)