lib: Allow framegroupkeys to return alternates
In some cases where subtly different data sets are reusing a single
class (eg ACSIS and DAS data in ACSIS format) we need to use different
keywords for determining whether two files come from the same
subsystem. If framegroupkeys() returns a reference to an array
in the final element the keywords in that array are gone through
until a keyword is present in the header.