acsis: Get things working with DAS data
gsd2acsis output looks like ACSIS files so we reduce as if they
are ACSIS data. Unfortunately there are some tweaks. In particular
the IFFREQ for DAS data is not fixed when sub-band merging is
required and instead we use a SPECID FITS header.

Grouping concatenates all IFFREQs.

Note that the input files have a rawfixedpart() of "h" but the
output files match the standard ACSIS convention of "a".