SCUBA2: only use amplitudes greater than 3-sigma
In observations of weak sources (primarily at 450 um), the fits to
extreme SMU positions may yield negative and/or weakly positive
amplitudes. While negative ones are immediately excluded, care must be
taken with the positive ones so that only statistically-significant
amplitudes are included. A simple 3-sigma cut has been implemented
here, and works well for the example of IRC+10216 data on 20120125
(obs 24).