SCUBA2: dummy QL/SUMMIT recipes for specific recipe choices
Users can choose the most suitable data processing recipe, which is
fine for offline reduction, but when running at the telescope their
choices should be overridden to fall back on standard QL and SUMMIT
recipes. Currently the pipeline does not have that capability, so
create symbolic links pointing to the standard QL/SUMMIT recipes as a
stopgap measure.