SCUBA2: use mean ERR in image if variance at peak is bad
Previously if the variance at the peak was bad, the primitive used the
average in a 3x3 region centred on the peak. Sometimes even this
results in bad values, so just use the whole image instead. This is
probably OK because the images have been trimmed to only 90" on a side
and the variance probably won't vary significantly across a map this
size.

As a last resort if the uncertainty is still undefined (something went
really badly wrong...), just set it to 1/10 of the peak value.