Index: substitute whitespaces with string when writing to index files
If any of the entries contains a space, ORAC-DR will abort with an
error, as it changes the apparent number of columns in the index file.

To avoid this, substitute whitespace with a "space" string (_SPACE_)
for writing. The string is removed from the entry before making
comparisons. Hopefully there won't be any entries that contain
"_SPACE_" already.