SCUBA2: skip noise calculation if we know the flatfield is bad
When retrieving the fastflats for setup observations the pipeline also
retrieves the number of bolometers. We can use that info to see if any
of the flatfield solutions failed, and if so skip the corresponding
noise calculation rather than wait for calcnoise to exit with bad
status.
(cherry picked from commit f4d621b01896a9a0665aae887542fe40dd9b45b5)