SCUBA2: check for NEWIMAGE flag before proceeding
Sometimes the summit pipeline contains data from a single subarray,
which will be in the form of flatfielded timeseries data when this
primitive is called. Non-image data trips up the NEFD calculation
because it assumes the presence of the exp_time NDF component, which
clearly won't exist if an image hasn't been made.

To avoid this, put in a check for the NEWIMAGE flag which indicates
that a (new) image has been created during this pass through the
recipe.