SCUBA2: use a good Frame object as a template in the QL
Sometimes data from the final SMU position is skipped in the QL
pipeline, which leads to problems downstream when the recipe attempts
to create the focus cube because some primitives implicitly assume the
current Frame contains a processed image. In this case the Frame
actually contains fastflat data and some important header items will
be missing. A simple solution is to use one of the processed Frame
objects as the current Frame, as all fastflat processing will have
been completed at this stage.