SCUBA2: proceed with SETUP in SUMMIT pipeline for mixed sequence data
If the SUMMIT pipeline has fallen behind, the Frame for a SETUP
observation will contain mixed sequence data from both the fastflat
and the noise. Processing the fastflat will set the PROCESS_FASTFLAT
flag, so trap this case and proceed with the noise calculation.

If the PROCESS_FASTFLAT flag is set, proceed only if we have mixed
sequence data for SETUPS in the SUMMIT pipeline.