SCUBA2: only flatfield data if necessary if making a map
If the pipeline has fallen behind and it will make a map this time
through the recipe, check to see if the pipeline needs to bother
flatfielding the data before making a map. Flatfielding can be avoided
if this is the first Frame in the Group, or if a Group image was
created on the previous pass through the recipe.