SCUBA2: unset the OBSCOMP parameter if the scan pattern has been completed
In some cases, the percentage completion condition can be met and
acted upon for an observation which repeats the scan pattern multiple
times. This can lead to the remaining data being ignored, leaving a
map made with only small fraction of the available data. So check to
see if the TCS_INDEX has increased and if so, unset the OBSCOMP
parameter to allow subsequent data to be processed.