SCUBA2: omit amplitudes that have changed sign from fit
Now the amplitudes are stored until the end of processing before they
are assessed for including in the fit. An amplitude is left out if it
is negative, less than 3-sigma or if it has changed sign compared with
the previous amplitude in a sense that will bias the fit.

In the latter situation there are two cases to deal with:

 - when the SMU position is changing towards the mid-point, if the
   current amplitude is negative and the previous one positive, the
   previous amplitude is ignored.

 - when the SMU position is changing away from the mid-point, if the
   current amplitude is positive and the previous one negative, the
   current amplitude is ignored.

The second of these cases has been tested, but not the first.
(cherry picked from commit ce2a038814b95ebd5432ca8c82acb13f63145219)