SCUBA2: remove BEAMEQUIV, calculate beam area to compare with empirical value
Since BEAMEQUIV was always mean to be a sanity check more than an FCF
to use, and now SCUBA-2 calibration is now better understood, there is
no need to calculate it. Instead we can use the derived ARCSEC and
BEAM FCFs to calculate the effective beam area (and an equivalent
FWHM) and compare that with the empirically-derived value as a measure
of the consistency of the respective FCFs.