SCUBA2: return the total number of good bolometers in NBOLO_SUM
A value will always be returned, but it will *always* be zero for raw
fastflats (TYPE=RAW or SETUP). If TYPE=SOL is selected, the number of
good bolometers for each solution will be totalled and returned as a
quick way of checking that at least 1 good solution was obtained.