SCUBA2: upgrade to handle the summit pipeline
The summit pipeline has different requirements than the QL in that it
sometimes requires data from other Frame objects in the Group, not
just the current Frame. In addition, in order to prevent filename
re-use (which reduces the amount of data used in the map-making step),
the list of fastflats must be at the end of the list of data files to
process.