SCUBA2: implement a more thorough validation of fitted amplitudes
The previous method of excluding negative amplitudes or those less
than 3-sigma wasn't implemented as thoroughly as intended. now use a
hash to track which amplitudes are "good" and "bad" and only create
the array of points to be fitted after all the SMU positions have been
analyzed.

Also the results are printed to the screen, with good values shown in
green (bad in cyan).