SCUBA2: print bad amplitudes as NaN, skip assessment of bad amplitudes
Occasionally maps of Mars end up with blank pixels at the centre
position. In these cases the bad values should not be printed out in
full when reporting the amplitudes so reformat as NaN. In addition, no
assessment of whether this is a good amplitude can be made - in fact
any attempt to do so could wrongly influence the assessment of
amplitudes from neighbouring SMU positions, so skip the appropriate
SMU position if there are blank pixels at the map centre.