SCUBA2: report total number of bolometers and average per subarray
As a way of tracking the health of the arrays, this primitive now
reports the total number of good bolometers in the most recent
flatfield and calculate the average number of good bolometers per
subarray. Perhaps this is a useful single number to track.
(cherry picked from commit 992120e6a1ef0884a16bf89483e52daa7ae75419)