SCUBA2: remove code meant for only processing fastflats
Way back in the day, when fastflats were first introduced I added this
code to process them when running offline to get a handle on how
stable the flatfields were. The intention was to ONLY process the
fastflats and ignore the science data altogether.

However, now the pipeline processes fastflats every time (rather than
storing the first batch for later), and the grouping for focus
observations means that the logic in the removed block results in
failures for moving sources. (Actually, the code does exactly what it
was originally intended to do, but only works for moving sources - in
the current scheme, however, it causes problems.)

If ever offline fastflat-only processing becomes a high priority, a
new method will have to be devised.