SCUBA2: store a blank QA report to avoid bad flatfields being skipped
The QA analysis has to report that subarrays with bad flatfields have
failed, but _QA_NOISE_ had no way to access that info because
_CALCULATE_NOISE_ only runs on data with good flatfields.

The simplest solution is to initialize the hash with the QA reporting
info before the noise calculation (rather than at the analysis
stage). _CALCULATE_NOISE_ now stores zeroes for all the QA and logging
parameters so _QA_NOISE_ can report flatfield failures.