scuba2: Force read of header to read NBOLOEFF
Prior to e10d33c083b5b7338217a42117013e297efe04a9 a call to
Frm->file() automatically triggered a readhdr() call. We now
don't do that (it was only happening for JSA instruments) so
we have to call it explicitly.

We may revisit the automatic readhdr since it's not
necessarily a bad idea.