SCUBA2: delay display of noise until after writing setup flag file
The display of the noise data involves creating a focal-plane mosaic,
which becomes the Frame file on exit from _DISPLAY_NOISE_. This file
then gets written to the setup flag file which causes problems for the
POINTING_FOCUS task as it's not in the index file. To get round this
and ensure that the names of the individual noise files are written to
the flag file, simply delay the display until after the flag file is
written.