SCUBA2: keep source and fit profiles on disk if requested
The x and y profiles are given suffices "_xprofile" and "_yprofile"
respectively. If the profile is of a matched-filtered map then the "x"
and "y" have "mf" appended.

The profile of the fit is copied into an NDF extension called
".more.smurf.fitprofile". That's probably not the right place but
people are probably used to seeing that path for other NDFs.